# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for managing categories, products, transactions, and POS operations\n\n## Features:\n- Category Management (CRUD)\n- Product Management (CRUD with category relationship)\n- Product Search by Name (case-insensitive partial match)\n- Transaction / Checkout (multi-item checkout with stock deduction)\n- Sales Reports (daily summary, date range, best selling product)\n- Dashboard Statistics",
        "title": "Retail Core API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/checkout": {
            "post": {
                "description": "Process a checkout with multiple items. Validates product availability, deducts stock, and creates a transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Process checkout",
                "parameters": [
                    {
                        "description": "Checkout request with items and quantities",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Checkout successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/report": {
            "get": {
                "description": "Retrieve the sales summary for a specific date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get sales report by date range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved report",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SalesReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing start_date or end_date",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/transactions": {
            "get": {
                "description": "Retrieve a paginated list of all transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transactions",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PaginatedTransactions"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "description": "Retrieve details of a specific transaction including its items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get a transaction by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid transaction ID",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helpers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation detail"
                },
                "message": {
                    "type": "string",
                    "example": "Error occurred"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "helpers.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.BestSellingProduct": {
            "description": "Best selling product information",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Indomie Goreng"
                },
                "qty_sold": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "models.CheckoutItem": {
            "description": "Single item to be checked out",
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer",
                    "example": 3
                },
                "quantity": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.CheckoutRequest": {
            "description": "Request body for processing a checkout",
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CheckoutItem"
                    }
                }
            }
        },
...
        "models.PaginatedTransactions": {
            "description": "Paginated list of transactions",
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionListItem"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "models.SalesReport": {
            "description": "Sales summary report with revenue, transaction count, and best seller",
            "type": "object",
            "properties": {
                "best_selling_product": {
                    "$ref": "#/definitions/models.BestSellingProduct"
                },
                "total_revenue": {
                    "type": "integer",
                    "example": 45000
                },
                "total_transactions": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.Transaction": {
            "description": "Transaction information with details of purchased items",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-08T12:00:00Z"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionDetail"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "total_amount": {
                    "type": "integer",
                    "example": 45000
                }
            }
        },
        "models.TransactionDetail": {
            "description": "Detail of a single item within a transaction",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "product_id": {
                    "type": "integer",
                    "example": 3
                },
                "product_name": {
                    "type": "string",
                    "example": "Indomie Goreng"
                },
                "quantity": {
                    "type": "integer",
                    "example": 5
                },
                "subtotal": {
                    "type": "integer",
                    "example": 15000
                },
                "transaction_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.TransactionListItem": {
            "description": "Transaction summary for list display",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-08T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "item_count": {
                    "type": "integer",
                    "example": 3
                },
                "total_amount": {
                    "type": "integer",
                    "example": 45000
                }
            }
        }
    }
}